Got 2 out of the 4 functions, making some moves here.

+++ cache_functions.h
@@ -1,6 +1,6 @@
 long block_size();
 long cache_size();
-long speed_main_memory();
-long speed_cache_memory();
+double speed_main_memory();
+double speed_cache_memory();
 double speed_array_accesses_sequenced();
 double speed_array_accesses_random();
